Update heat.ino
Problem with watchdog solved (untested yet)

+++ heat.ino
@@ -1,10 +1,10 @@
 #include "heat.h"
 
-const float baseClock = 200;                            //200 Millisekunden
-const float level1Divider = 1000/baseClock;             //1 Sekunde
-const float level2Divider = 2000/baseClock;             //2 Sekunden
-const float level3Divider = 60000/baseClock;            //1 Minute
-const float level4Divider = 600000/baseClock;           //10 Minuten
+const unsigned long baseClock = 200;                            //200 Millisekunden
+const unsigned long level1Divider = 1000/baseClock;             //1 Sekunde
+const unsigned long level2Divider = 2000/baseClock;             //2 Sekunden
+const unsigned long level3Divider = 60000/baseClock;            //1 Minute
+const unsigned long level4Divider = 600000/baseClock;           //10 Minuten
 
 unsigned long cycleCounter;
 unsigned long lastTime;
@@ -49,7 +49,7 @@ void loop() {
     Controller::output();
 
     //Level1 - jede Sekunde                                             //Offset: 0
-    if ((cycleCounter + 0) % ((unsigned long) level1Divider) == 0) {
+    if ((cycleCounter + 0) % level1Divider == 0) {
       RealTimeClock::addSecond();
       Lcd::refresh();
       #if defined(NO_PERIPHERY) 
@@ -58,7 +58,7 @@ void loop() {
     }
 
     //Level2 - alle 2 Sekunden                                          //Offset: 1
-    if ((cycleCounter + 1) % ((unsigned long) level2Divider) == 0) {
+    if ((cycleCounter + 1) % level2Divider == 0) {
       TemperatureSensors::refreshTemperatureValues();
       TemperatureSensors::startTemperatureConversion();
       Controller::compute();
@@ -68,7 +68,7 @@ void loop() {
     }
 
     //Level2 - alle 2 Sekunden                                          //Offset: 2
-    if ((cycleCounter + 2)  % ((unsigned long) level2Divider) == 0) {
+    if ((cycleCounter + 2)  % level2Divider == 0) {
       Rs485::communicate();                                                                   //----------------------voruebergehend hierher verschoben
       Error::output();
       Lcd::toggleValues();
@@ -81,7 +81,7 @@ void loop() {
     }
 
     //Level3 - jede Minute                                              //Offset: 3
-    if ((cycleCounter + 3)  % ((unsigned long) level3Divider) == 0) {
+    if ((cycleCounter + 3)  % level3Divider == 0) {
       RealTimeClock::refresh();
       #if defined(NO_PERIPHERY) 
       Serial.println("Level3-1"); 
@@ -89,7 +89,7 @@ void loop() {
     }
 
     //Level3 - jede Minute                                              //Offset: 4  (Achtung: 4 ist der maximale Offset. Bei 5 gibt es schon Überschneidungen mit dem Level 1.)
-    if ((cycleCounter + 4) % ((unsigned long) level3Divider) == 0) {
+    if ((cycleCounter + 4) % level3Divider == 0) {
       Logger::writeToSD();
       #if defined(NO_PERIPHERY) 
       Serial.println("Level3-2"); 
@@ -97,7 +97,7 @@ void loop() {
     }
 
     //Level4 - alle 10 Minuten
-    if ((cycleCounter + 0) % ((unsigned long) level4Divider) == 0) {   //Offset: 0
+    if ((cycleCounter + 0) % level4Divider == 0) {   //Offset: 0
       TemperatureSensors::requestOutsideTemperature();
       #if defined(NO_PERIPHERY) 
       Serial.println("Level4");
